refactor: 將所有跟金錢相關的型態都改為decimal

+++ Shared/Domain/Auction.cs
@@ -6,12 +6,12 @@ public class Auction
 {
     private LandContract landContract;
     private Player? highestBidder;
-    private int highestPrice;
+    private decimal highestPrice;
 
     public Auction(LandContract landContract)
     {
         this.landContract = landContract;
-        highestPrice = (int)(landContract.Land.Price * 0.5);
+        highestPrice = (decimal)(((int)landContract.Land.Price) * 0.5);
     }
     /// <summary>
     /// 結算拍賣
@@ -31,7 +31,7 @@ public class Auction
         }
         else // 流拍
         {
-            landContract.Owner.AddMoney((int)(landContract.Land.Price * 0.7));
+            landContract.Owner.AddMoney((decimal)(((int)landContract.Land.Price) * 0.7));
         }
     }
